Create meatloaf.html and update index.html's title
Add a recipes directory under the repository root in which to place all
recipe pages, and add a page for "The Best Meatloaf Recipe Ever" recipe
to that new directory.

Give the meatloaf.html page an h1 heading with the recipe name,
set the title element with the site name and recipe shortname, and
udpate the title element of index.html to be consistent with the
convention implemented in meatloaf.html's title element.